% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{bagging_pls}
\alias{bagging_pls}
\title{Bagging PLS — formula entry point.}
\usage{
bagging_pls(
  formula,
  data,
  ncomp = 2L,
  n_estimators = 50L,
  seed = 0L,
  na.action = stats::na.omit
)
}
\description{
Bagging PLS — formula entry point.
}
