% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{bagging_pls_fit}
\alias{bagging_pls_fit}
\title{Bagging PLS (bootstrap aggregation of PLS regressors).}
\usage{
bagging_pls_fit(X, Y, n_components, n_estimators = 50L, seed = 0L)
}
\description{
Bagging PLS (bootstrap aggregation of PLS regressors).
}
