% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{boosting_pls}
\alias{boosting_pls}
\title{Boosting PLS — formula entry point.}
\usage{
boosting_pls(
  formula,
  data,
  ncomp = 2L,
  n_estimators = 50L,
  learning_rate = 0.1,
  na.action = stats::na.omit
)
}
\description{
Boosting PLS — formula entry point.
}
