% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{boosting_pls_fit}
\alias{boosting_pls_fit}
\title{Boosting PLS (stage-wise refit with learning_rate).}
\usage{
boosting_pls_fit(X, Y, n_components, n_estimators = 50L, learning_rate = 0.1)
}
\description{
Boosting PLS (stage-wise refit with learning_rate).
}
