% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{continuum_regression}
\alias{continuum_regression}
\title{Continuum regression — formula entry point.}
\usage{
continuum_regression(
  formula,
  data,
  ncomp = 2L,
  tau = 0.5,
  na.action = stats::na.omit
)
}
\description{
Continuum regression — formula entry point.
}
