% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{continuum_regression_fit}
\alias{continuum_regression_fit}
\title{Continuum regression (tau ∈ [0, 1]).}
\usage{
continuum_regression_fit(X, Y, n_components, tau = 0.5)
}
\description{
Continuum regression (tau ∈ [0, 1]).
}
