% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_methods.R
\name{cppls}
\alias{cppls}
\title{Canonical Powered PLS — formula entry point.}
\usage{
cppls(formula, data, ncomp = 2L, gamma = 0.5, na.action = stats::na.omit)
}
\description{
Canonical Powered PLS — formula entry point.
}
