% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{di_pls_fit}
\alias{di_pls_fit}
\title{Domain-Invariant PLS (Nikzad-Langerodi 2018).}
\usage{
di_pls_fit(X_source, Y_source, n_components, X_target, di_lambda = 1)
}
\description{
Domain-Invariant PLS (Nikzad-Langerodi 2018).
}
