% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{ds_fit}
\alias{ds_fit}
\title{Direct Standardization (calibration transfer).}
\usage{
ds_fit(X_source, X_target)
}
\arguments{
\item{X_source}{Method-specific parameter. See the underlying `*_fit()` function for the exact semantics.}

\item{X_target}{Method-specific parameter. See the underlying `*_fit()` function for the exact semantics.}
}
\description{
Direct Standardization (calibration transfer).
}
