% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{ecr}
\alias{ecr}
\title{Elastic Component Regression — formula entry point.}
\usage{
ecr(formula, data, ncomp = 2L, alpha = 0.5, na.action = stats::na.omit)
}
\description{
Elastic Component Regression — formula entry point.
}
