% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{fused_sparse_pls_fit}
\alias{fused_sparse_pls_fit}
\title{Fused-sparse PLS (L1 + adjacent-coef smoothing).}
\usage{
fused_sparse_pls_fit(
  X,
  Y,
  n_components,
  l1_lambda = 0.05,
  fusion_lambda = 0.05
)
}
\description{
Fused-sparse PLS (L1 + adjacent-coef smoothing).
}
