% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{kernel_pls_fit}
\alias{kernel_pls_fit}
\title{Non-linear kernel PLS (Rosipal & Trejo 2001).}
\usage{
kernel_pls_fit(
  X,
  Y,
  n_components,
  kernel_type = 1L,
  gamma = 0,
  coef0 = 1,
  degree = 3L
)
}
\description{
Non-linear kernel PLS (Rosipal & Trejo 2001).
}
