% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{lw_pls_fit}
\alias{lw_pls_fit}
\title{Locally-weighted PLS (Næs & Centner 1998).}
\usage{
lw_pls_fit(X, Y, n_components, n_neighbors = 30L)
}
\description{
Locally-weighted PLS (Næs & Centner 1998).
}
