% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{missing_aware_nipals_fit}
\alias{missing_aware_nipals_fit}
\title{Missing-aware NIPALS PLS (Nelson 1996).}
\usage{
missing_aware_nipals_fit(X, Y, n_components)
}
\arguments{
\item{X}{Numeric matrix of predictors (rows = samples, cols = features).}

\item{Y}{Numeric matrix or vector of responses, with one row per sample.}

\item{n_components}{Integer. Number of latent components.}
}
\description{
Missing-aware NIPALS PLS (Nelson 1996).
}
