% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{on_pls_fit}
\alias{on_pls_fit}
\title{OnPLS — Orthogonal multi-block PLS (joint + unique loadings).}
\usage{
on_pls_fit(X, Y, n_joint, n_unique_per_block, block_sizes)
}
\description{
OnPLS — Orthogonal multi-block PLS (joint + unique loadings).
}
