% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{one_se_rule}
\alias{one_se_rule}
\title{One-SE rule from a (max_components × n_folds) fold RMSE matrix.}
\usage{
one_se_rule(fold_rmse_matrix)
}
\description{
One-SE rule from a (max_components × n_folds) fold RMSE matrix.
}
