% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pls4all_predict}
\alias{pls4all_predict}
\title{Predict with a fitted pls4all model.}
\usage{
pls4all_predict(model, X)
}
\arguments{
\item{model}{External pointer returned by [pls4all_fit()].}

\item{X}{Numeric matrix, n_new x p.}
}
\value{
Numeric matrix, n_new x n_targets.
}
\description{
Predict with a fitted pls4all model.
}
