% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{n4m_version}
\alias{n4m_version}
\title{Runtime version string of the loaded libn4m.}
\usage{
n4m_version()
}
\value{
A character scalar like "1.0.2+abi.2.0.0".
}
\description{
Runtime version string of the loaded libn4m.
}
