% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{pls_lda_fit}
\alias{pls_lda_fit}
\title{PLS-LDA — Linear Discriminant Analysis on PLS scores.}
\usage{
pls_lda_fit(X, y_labels, n_components, n_classes = NULL)
}
\description{
PLS-LDA — Linear Discriminant Analysis on PLS scores.
}
