% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{pls_logistic_fit}
\alias{pls_logistic_fit}
\title{Multinomial logistic regression on PLS scores.}
\usage{
pls_logistic_fit(X, y_labels, n_components, n_classes = NULL)
}
\description{
Multinomial logistic regression on PLS scores.
}
