% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{pls_qda_fit}
\alias{pls_qda_fit}
\title{PLS-QDA (Quadratic Discriminant Analysis on PLS scores).}
\usage{
pls_qda_fit(X, y_labels, n_components)
}
\description{
PLS-QDA (Quadratic Discriminant Analysis on PLS scores).
}
