% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_methods.R
\name{predict.pls4all_method_fit}
\alias{predict.pls4all_method_fit}
\title{Predict from a MethodResult-based pls4all fit.}
\usage{
\method{predict}{pls4all_method_fit}(object, newdata = NULL, ...)
}
\description{
Predict from a MethodResult-based pls4all fit.
}
