% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{random_frog_select}
\alias{random_frog_select}
\title{Random Frog (Phase 5g).}
\usage{
random_frog_select(
  X,
  Y,
  n_components,
  n_iterations = 100L,
  initial_size = 30L,
  min_size = NULL,
  max_size = NULL,
  top_k = 10L,
  seed = 0L
)
}
\description{
Random Frog (Phase 5g).
}
