% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{random_subspace_pls_fit}
\alias{random_subspace_pls_fit}
\title{Random-subspace PLS (Ho 1998).}
\usage{
random_subspace_pls_fit(
  X,
  Y,
  n_components,
  n_estimators = 50L,
  features_per_subspace = 10L,
  seed = 0L
)
}
\description{
Random-subspace PLS (Ho 1998).
}
