% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{ridge_pls_fit}
\alias{ridge_pls_fit}
\title{L2-augmented PLS regression.}
\usage{
ridge_pls_fit(X, Y, n_components, ridge_lambda = 1)
}
\arguments{
\item{X}{Numeric matrix of predictors (rows = samples, cols = features).}

\item{Y}{Numeric matrix or vector of responses, with one row per sample.}

\item{n_components}{Integer. Number of latent components.}

\item{ridge_lambda}{Numeric >= 0. Ridge regularisation strength.}
}
\description{
L2-augmented PLS regression.
}
