% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{rosa_fit}
\alias{rosa_fit}
\title{ROSA — Response-Oriented Sequential Alternation.}
\usage{
rosa_fit(X, Y, n_components, block_sizes)
}
\description{
ROSA — Response-Oriented Sequential Alternation.
}
