% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{scars_select}
\alias{scars_select}
\title{SCARS — Stability + CARS.}
\usage{
scars_select(
  X,
  Y,
  n_components,
  n_iterations = 50L,
  min_features = 5L,
  sample_fraction = 0.8,
  seed = 0L
)
}
\description{
SCARS — Stability + CARS.
}
