% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{so_pls_fit}
\alias{so_pls_fit}
\title{Sequential & Orthogonalised multi-block PLS (Næs et al. 2011).
`block_sizes` integer vector summing to ncol(X);
`n_components_per_block` integer vector of same length.}
\usage{
so_pls_fit(X, Y, block_sizes, n_components_per_block)
}
\description{
Sequential & Orthogonalised multi-block PLS (Næs et al. 2011).
`block_sizes` integer vector summing to ncol(X);
`n_components_per_block` integer vector of same length.
}
