% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_methods.R
\name{sparse_pls}
\alias{sparse_pls}
\title{Sparse SIMPLS — formula entry point.}
\usage{
sparse_pls(
  formula,
  data,
  ncomp = 2L,
  sparsity_lambda = 0.05,
  na.action = stats::na.omit
)
}
\description{
Sparse SIMPLS — formula entry point.
}
