% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{sparse_pls_da_fit}
\alias{sparse_pls_da_fit}
\title{Sparse PLS-DA classifier (`y_labels` is an integer vector of class IDs).}
\usage{
sparse_pls_da_fit(X, y_labels, n_components, sparsity_lambda = 0.05)
}
\description{
Sparse PLS-DA classifier (`y_labels` is an integer vector of class IDs).
}
