% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{uve_select}
\alias{uve_select}
\title{UVE — Uninformative Variable Elimination.}
\usage{
uve_select(X, Y, n_components, noise_features = NULL, noise_seed = 0L)
}
\description{
UVE — Uninformative Variable Elimination.
}
