% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{bipls_select}
\alias{bipls_select}
\title{biPLS — backward interval PLS.}
\usage{
bipls_select(X, Y, n_components, interval_width = 10L, min_intervals = 1L)
}
\description{
biPLS — backward interval PLS.
}
