% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{emcuve_select}
\alias{emcuve_select}
\title{EMCUVE — ensemble Monte Carlo UVE.}
\usage{
emcuve_select(
  X,
  Y,
  n_components,
  noise_features = NULL,
  noise_seed = 0L,
  n_ensembles = 5L,
  vote_threshold = 0.5
)
}
\description{
EMCUVE — ensemble Monte Carlo UVE.
}
