% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{ga_select}
\alias{ga_select}
\title{GA-PLS — genetic algorithm variable selection.}
\usage{
ga_select(
  X,
  Y,
  n_components,
  n_generations = 50L,
  population_size = 50L,
  min_features = NULL,
  max_features = NULL,
  mutation_rate = 0.01,
  seed = 0L
)
}
\description{
GA-PLS — genetic algorithm variable selection.
}
