% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{gpr_pls_fit}
\alias{gpr_pls_fit}
\title{Gaussian Process Regression on PLS scores (single-target Y).}
\usage{
gpr_pls_fit(
  X,
  Y,
  n_components,
  length_scale = 1,
  noise_level = 1e-04,
  seed = 0L
)
}
\description{
Gaussian Process Regression on PLS scores (single-target Y).
}
