% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{group_sparse_pls_fit}
\alias{group_sparse_pls_fit}
\title{Group-sparse PLS (group L1 across feature groups).}
\usage{
group_sparse_pls_fit(X, Y, n_components, group_assignment, group_lambda = 0.05)
}
\description{
Group-sparse PLS (group L1 across feature groups).
}
