% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{ipw_select}
\alias{ipw_select}
\title{IPW-PLS.}
\usage{
ipw_select(
  X,
  Y,
  n_components,
  n_iterations = 10L,
  top_k = 10L,
  damping = 0.5,
  weight_floor = 1e-06
)
}
\description{
IPW-PLS.
}
