% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{irf_select}
\alias{irf_select}
\title{IRF — Interval Random Frog.}
\usage{
irf_select(
  X,
  Y,
  n_components,
  n_iterations = 100L,
  window_size = 10L,
  initial_intervals = 10L,
  top_k = 5L,
  seed = 0L
)
}
\description{
IRF — Interval Random Frog.
}
