% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{iriv_select}
\alias{iriv_select}
\title{IRIV — Iteratively Retains Informative Variables.}
\usage{
iriv_select(X, Y, n_components, max_rounds = 20L, seed = 0L)
}
\description{
IRIV — Iteratively Retains Informative Variables.
}
