% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{n_pls_fit}
\alias{n_pls_fit}
\title{N-PLS (3-way tensor) regression. `X` is the flattened (n, mode_j*mode_k) matrix.}
\usage{
n_pls_fit(X_flat, Y, n_components, mode_j, mode_k)
}
\description{
N-PLS (3-way tensor) regression. `X` is the flattened (n, mode_j*mode_k) matrix.
}
