% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{o2pls_fit}
\alias{o2pls_fit}
\title{O2-PLS (bi-directional OPLS).}
\usage{
o2pls_fit(X, Y, n_predictive = 2L, n_x_orthogonal = 1L, n_y_orthogonal = 1L)
}
\description{
O2-PLS (bi-directional OPLS).
}
