% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{pds_fit}
\alias{pds_fit}
\title{Piecewise Direct Standardization (calibration transfer).}
\usage{
pds_fit(X_source, X_target, window_half_width = 2L)
}
\description{
Piecewise Direct Standardization (calibration transfer).
}
