% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{n4m_abi_version}
\alias{n4m_abi_version}
\title{Loaded ABI version as an integer vector (major, minor, patch).}
\usage{
n4m_abi_version()
}
\value{
An integer vector of length 3.
}
\description{
Loaded ABI version as an integer vector (major, minor, patch).
}
