% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{pls4all_version}
\alias{pls4all_version}
\title{Runtime version string of the loaded libp4a.}
\usage{
pls4all_version()
}
\value{
A character scalar like "0.67.0+abi.1.1.0".
}
\description{
Runtime version string of the loaded libp4a.
}
