% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{pso_select}
\alias{pso_select}
\title{PSO-PLS (Binary Particle Swarm Optimization).}
\usage{
pso_select(
  X,
  Y,
  n_components,
  n_swarm = 30L,
  n_iterations = 50L,
  w = 0.729,
  c1 = 1.494,
  c2 = 1.494,
  v_max = 4,
  seed = 0L
)
}
\description{
PSO-PLS (Binary Particle Swarm Optimization).
}
