% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{randomization_select}
\alias{randomization_select}
\title{Randomization test selector.}
\usage{
randomization_select(
  X,
  Y,
  n_components,
  n_permutations = 100L,
  randomization_seed = 0L,
  alpha = 0.05
)
}
\description{
Randomization test selector.
}
