% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{ridge_pls}
\alias{ridge_pls}
\title{Ridge PLS — formula entry point.}
\usage{
ridge_pls(
  formula,
  data,
  ncomp = 2L,
  ridge_lambda = 1,
  na.action = stats::na.omit
)
}
\description{
Ridge PLS — formula entry point.
}
