% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_extra.R
\name{robust_pls}
\alias{robust_pls}
\title{Robust PLS — formula entry point.}
\usage{
robust_pls(
  formula,
  data,
  ncomp = 2L,
  huber_k = 1.345,
  max_irls_iter = 20L,
  na.action = stats::na.omit
)
}
\description{
Robust PLS — formula entry point.
}
