% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{robust_pls_fit}
\alias{robust_pls_fit}
\title{Robust PLS via Huber IRLS.}
\usage{
robust_pls_fit(X, Y, n_components, huber_k = 1.345, max_irls_iter = 20L)
}
\description{
Robust PLS via Huber IRLS.
}
