% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{shaving_select}
\alias{shaving_select}
\title{Shaving selector.}
\usage{
shaving_select(
  X,
  Y,
  n_components,
  n_steps = 10L,
  min_features = 5L,
  shave_fraction = 0.1
)
}
\description{
Shaving selector.
}
