% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{sipls_select}
\alias{sipls_select}
\title{siPLS — synergistic interval PLS.}
\usage{
sipls_select(X, Y, n_components, interval_width = 10L, combination_size = 2L)
}
\description{
siPLS — synergistic interval PLS.
}
