% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{stability_select}
\alias{stability_select}
\title{Stability selector (coefficient stability, MCUVE-style).}
\usage{
stability_select(X, Y, n_components, top_k = 10L)
}
\description{
Stability selector (coefficient stability, MCUVE-style).
}
