% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{t2_select}
\alias{t2_select}
\title{T²-PLS — sweep over α-thresholds.}
\usage{
t2_select(X, Y, n_components, alpha_thresholds, min_selected = NULL)
}
\description{
T²-PLS — sweep over α-thresholds.
}
