% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{vissa_select}
\alias{vissa_select}
\title{VISSA — Variable Iterative Space Shrinkage Approach.}
\usage{
vissa_select(
  X,
  Y,
  n_components,
  n_iterations = 20L,
  n_submodels = 100L,
  ratio_kept = 0.1,
  threshold = 0.5,
  floor_probability = 0.01,
  seed = 0L
)
}
\description{
VISSA — Variable Iterative Space Shrinkage Approach.
}
