% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{wvc_select}
\alias{wvc_select}
\title{WVC-PLS — weighted vector correlation top-k selector.}
\usage{
wvc_select(X, Y, n_components, top_k = 10L, normalize = TRUE)
}
\description{
WVC-PLS — weighted vector correlation top-k selector.
}
