% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extra.R
\name{wvc_threshold_select}
\alias{wvc_threshold_select}
\title{WVC-threshold selector.}
\usage{
wvc_threshold_select(
  X,
  Y,
  n_components,
  normalize = TRUE,
  threshold = 0,
  threshold_factor = 1,
  min_selected = 1L
)
}
\description{
WVC-threshold selector.
}
